requirements.txt
Django==1.7
python-crontab==1.9.0
daemonocle==0.8
psycopg2==2.6.1
gunicorn==19.4.5
picamera==1.10
